Start mechanising overflow checks
This comes from me not being able to keep the overflow checks I need
for the constant expression generator in my head in one go.

The overflow checks currently mechanised probably tell me enough to
short out the constant generator when it's about to generate an
overflow, but not how to rescue the situation.