Add var capping to storelikes
The easiest way for now is to make all storelikes report a maximum
amount of variables they will declare, and bung that into the
availability check (if it's zero, don't bother checking; this is a new
change to the var capper).

This is where a unified availability/payload monad would come in handy:
every time you state that you want to generate a new variable, the
availability check would increase the cap requirement by one.