Add initial_value key to delitmus variables
This is just the same as the init mapping for globals, but takes on the
initialiser value for locals; it's also resilient to variable renames.

Fixes #248.

(Note that this does *not* convert to assignments, leaving the onus for
doing so on the user of the delitmus for now.  We might add this as a
flag later.)