Fix silly cmpxchg generating mistake
I thought that setting arbitrary cmpxchg multi-safe to `If_no_cycles
would be ok, but forgot that this causes the generator to reject every
single cmpxchg as we consider them as always having dependency cycles.
Oops.  This is the second time this has happened, so I've written myself
a warning.