Remove documentation tree
This has got so outdated that it's a liability at this stage;
people might actually try to use this documentation and then
wonder why all the commands are completely different.

As with Docker, I'm not saying there shouldn't be documentation;
I'm just pushing this back to 'we need to implement this properly
sometime'.